Organize package for pip install
