+++ coda/__init__.py
@@ -1,1 +1,3 @@
 from .coda import CODA
+from .datasets import Dataset
+from .oracle import Oracle
