+++ coda/coda.py
@@ -6,7 +6,7 @@ import random
 import mlflow
 
 from coda.base import ModelSelector
-from surrogates import Ensemble
+from .surrogates import Ensemble
 
 _DEBUG = True
 from logging_util import plot_bar
