+++ scripts/launch_all_methods.py
