+++ scripts/model_picker_epsilon_gridsearch.py
@@ -22,7 +22,7 @@ import numpy as np
 import torch
 from tqdm import tqdm
 
-from datasets import Dataset
+from coda.datasets import Dataset
 from coda.baselines import ModelPicker
 from main import seed_all
